Played a bit with the transform from the TestBehaviour

+++ src/Domain/Timer.cpp
@@ -1,6 +1,6 @@
 #include "Timer.h"
 
-Timer::Timer(std::function<void ()> func, int secDelay, int repeatCount, bool waitOneDelayTime) :
+Timer::Timer(std::function<void ()> func, float secDelay, int repeatCount, bool waitOneDelayTime) :
     func(func),
     secDelay(secDelay),
     repeatCount(repeatCount),
@@ -19,7 +19,7 @@ void Timer::TimerLoop() const
 
     while( !stop && (repeatCount == -1 || count < repeatCount) )
     {
-        std::this_thread::sleep_for(std::chrono::milliseconds(secDelay * 1000));
+        std::this_thread::sleep_for(std::chrono::milliseconds(int(secDelay * 1000)));
         func();
         ++count;
     }
